Lorenz Attractor with 3d rot libs
Lorenz Attractor with 3d rot libs.
Next: Add color to Lorenz Attractor scene and move on.
Add rotating wiremeshes.

+++ L1Demo_2018_0x24D.X/alu.h
@@ -3,14 +3,14 @@
 
 #define M_PI 3.141592653589
 
-extern float rotationMatrix[4][4];
-extern float inputMatrix[4][1];
-extern float outputMatrix[4][1]; 
+extern double rotationMatrix[4][4];
+extern double inputMatrix[4][1];
+extern double outputMatrix[4][1]; 
 
 float radians(uint16_t angle);
 int realtoint(float oldval, float oldmin, float oldmax, float newmin, float newmax);
 
-void alu_rot(float x, float y, float z);
-void alu_calc_rot_matrix(float angle, float u, float v, float w);
+void alu_rot(double x, double y, double z);
+void alu_calc_rot_matrix(double angle, double u, double v, double w);
 
 #endif
